Update generator DLL path in project file
Updated the `<None>` item in the project file to reference the `IeuanWalker.MinimalApi.Endpoints.Generator.dll` file from the `bin\$(Configuration)\netstandard2.0\` directory. This replaces the previous reference to the `..\MinimalApi.Generator\` directory, ensuring the project uses the correct built output of the generator. The change supports dynamic adaptation to the current build configuration.