Update project structure and add analyzer reference
Removed the explicit inclusion of the `GlobalExceptionHandler\`
folder from the project file as it is no longer needed. Added
a reference to `IeuanWalker.MinimalApi.Endpoints.Generator.dll`
from the `bin\$(Configuration)\netstandard2.0\` directory. The
DLL is configured to be packed into the NuGet package under
`analyzers/dotnet/cs` and marked as invisible in the project.